added package dependencies, added screen flow between firebase authentication and weather app
